feat: new back and close buttons used for react-navigation header
Keeping the legacy version of the close button, with a new name.
Updated component previously using `HeaderRightClose`
