+++ apps/ledger-live-mobile/src/components/RootNavigator/WalletConnectNavigator.tsx
@@ -8,7 +8,7 @@ import TransparentHeaderNavigationOptions from "../../navigation/TransparentHead
 import WalletConnectScan from "../../screens/WalletConnect/Scan";
 import WalletConnectConnect from "../../screens/WalletConnect/Connect";
 import WalletConnectDeeplinkingSelectAccount from "../../screens/WalletConnect/DeeplinkingSelectAccount";
-import HeaderRightClose from "../HeaderRightClose";
+import { NavigationHeaderCloseButtonAdvanced } from "../NavigationHeaderCloseButton";
 import { getStackNavigatorConfig } from "../../navigation/navigatorConfig";
 import { WalletConnectNavigatorParamList } from "./types/WalletConnectNavigator";
 
@@ -30,7 +30,10 @@ export default function WalletConnectNavigator() {
           ...TransparentHeaderNavigationOptions,
           title: "Wallet Connect",
           headerRight: () => (
-            <HeaderRightClose color={colors.white} preferDismiss={false} />
+            <NavigationHeaderCloseButtonAdvanced
+              color={colors.white}
+              preferDismiss={false}
+            />
           ),
           headerLeft: () => null,
         }}
@@ -40,7 +43,9 @@ export default function WalletConnectNavigator() {
         component={WalletConnectDeeplinkingSelectAccount}
         options={{
           title: t("walletconnect.deeplinkingTitle"),
-          headerRight: () => <HeaderRightClose preferDismiss={false} />,
+          headerRight: () => (
+            <NavigationHeaderCloseButtonAdvanced preferDismiss={false} />
+          ),
           headerLeft: () => null,
         }}
       />
