+++ apps/ledger-live-mobile/src/components/WebPTXPlayer/index.tsx
@@ -30,7 +30,7 @@ import {
 } from "../RootNavigator/types/helpers";
 import { BaseNavigatorStackParamList } from "../RootNavigator/types/BaseNavigator";
 import { initialWebviewState } from "../Web3AppWebview/helpers";
-import HeaderRightClose from "../HeaderRightClose";
+import { NavigationHeaderCloseButtonAdvanced } from "../NavigationHeaderCloseButton";
 
 type BackToWhitelistedDomainProps = {
   manifest: AppManifest;
@@ -62,7 +62,7 @@ function BackToWhitelistedDomain({ manifest }: BackToWhitelistedDomainProps) {
 function HeaderRight() {
   const { colors } = useTheme();
 
-  return <HeaderRightClose color={colors.neutral.c100} />;
+  return <NavigationHeaderCloseButtonAdvanced color={colors.neutral.c100} />;
 }
 
 type Props = {
